% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{as_group}
\alias{as_group}
\alias{as_group.grob}
\alias{as_group.Layer}
\title{Collect layers into a group that can be treated as a single layer}
\usage{
as_group(..., id)

\method{as_group}{grob}(..., id)

\method{as_group}{Layer}(..., id)
}
\arguments{
\item{...}{A range of layers to combine}

\item{id}{An id that can be used to reference this filter somewhere else}
}
\value{
A list of layers oor a \link[grid:grid.grob]{gTree} depending on the input
}
\description{
While you often want to apply filters to layers one by one, there are times
when one filter should be applied to a collection of layers as if they were
one. This can be achieved by first combining all the layers into a group with
\code{as_group()} and applying the filter to the resulting group. This can only be
done to ggplot2 layers and grobs as the other supported objects are not part
of a graphic stack.
}
