% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{Channels}
\alias{Channels}
\alias{ch_red}
\alias{ch_green}
\alias{ch_blue}
\alias{ch_alpha}
\alias{ch_hue}
\alias{ch_chroma}
\alias{ch_luminance}
\alias{ch_saturation}
\alias{ch_lightness}
\title{Set a channel of interest from a layer}
\usage{
ch_red(x)

ch_green(x)

ch_blue(x)

ch_alpha(x)

ch_hue(x)

ch_chroma(x)

ch_luminance(x)

ch_saturation(x)

ch_lightness(x)
}
\arguments{
\item{x}{Any object interpretable as a layer}
}
\value{
\code{x} with a channel spec attached
}
\description{
Some effects uses a particular channel for specific parameters, such as
\code{\link[=with_displace]{with_displace()}}, which grabs the relative x and y displacements from
different channels in some other layer. To facilitate specifying which
channel to use from a layer (which is always multichannel), you can wrap the
specification in a channel specifier given below. If a filter requires a
specific channel and none is specified it will default to \code{luminance} (based
on the \code{hcl} colour space)
}
