% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfx-package.R
\docType{package}
\name{ggfx-package}
\alias{ggfx}
\alias{ggfx-package}
\title{ggfx: Filters and Shaders for 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a range of filters and shaders that can be applied to 
    layers from the 'ggplot2' package and its extensions. The filters are 
    applied at draw time and will thus respond to resizing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/ggfx}
  \item Report bugs at \url{https://github.com/thomasp85/ggfx/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
