% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-helpers.R
\name{render_context}
\alias{render_context}
\alias{viewport_location}
\alias{viewport_is_clipping}
\alias{current_resolution}
\alias{as_pixels}
\alias{from_pixels}
\title{Rendering information}
\usage{
viewport_location()

viewport_is_clipping()

current_resolution()

as_pixels(x)

from_pixels(x)
}
\arguments{
\item{x}{A numeric or \link[grid:unit]{grid::unit} object}
}
\value{
See details
}
\description{
These utility functions can help when creating custom filters (using
\code{\link[=with_custom]{with_custom()}}) as they can provide information about the current rendering
context.
}
\details{
\itemize{
\item \code{viewport_location()}: Returns the bounding box defining the current
viewport in pixels in the order \code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}
\item \code{viewport_is_clipping()}: Returns \code{TRUE} if the current viewport has
clipping turned on
\item \code{current_resolution()}: Returns the resolution of the active device in ppi
(pixels-per-inch)
\item \code{as_pixels(x)}: Converts \code{x} to pixels if \code{x} is given as a unit object. It
is assumed that x encodes a dimension and not a location. If \code{x} is a
numeric it is assumed to already be in pixels
\item \code{from_pixels}: Converts a numeric giving some pixel dimension to a unit
object.
}
}
\keyword{internal}
