% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{with_blur}
\alias{with_blur}
\alias{with_blur.grob}
\alias{with_blur.Layer}
\alias{with_blur.ggplot}
\title{Apply a gaussian blur to your layer}
\usage{
with_blur(x, sigma = 0.5, radius = 0, stack = FALSE, ...)

\method{with_blur}{grob}(x, sigma, radius, stack = FALSE,
  background = NULL, ...)

\method{with_blur}{Layer}(x, sigma = 0.5, radius = 0, stack = FALSE,
  ...)

\method{with_blur}{ggplot}(x, sigma = 0.5, radius = 0, stack = FALSE,
  ignore_background = TRUE, ...)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, or a grob}

\item{sigma}{The standard deviation of the gaussian kernel. Increase it to
apply more blurring}

\item{radius}{The radius of the blur. If \code{0} it will be calculated
automatically from \code{sigma}}

\item{stack}{Should the original layer be placed on top?}

\item{...}{Arguments to be passed on to methods}

\item{background}{A grob to draw below the filtered grob.}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter adds a blur to the provided ggplot layer. The amount of blur can
be controlled and the result can optionally be put underneath the original
layer.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_blur(geom_point(data = mtcars, size = 3), sigma = 3)

}
