% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displace.R
\name{with_displacement}
\alias{with_displacement}
\alias{with_displacement.grob}
\alias{with_displacement.Layer}
\alias{with_displacement.ggplot}
\title{Apply a displacement map to a layer}
\usage{
with_displacement(
  x,
  map,
  x_channel,
  y_channel,
  scale,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{grob}(
  x,
  map,
  x_channel,
  y_channel,
  scale,
  ...,
  background = NULL,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{Layer}(
  x,
  map,
  x_channel,
  y_channel,
  scale,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{ggplot}(
  x,
  map,
  x_channel,
  y_channel,
  scale,
  ignore_background = TRUE,
  ...,
  id = NULL,
  include = is.null(id)
)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, or a grob}

\item{map}{The displacement map to use. Can either be a string identifying a
registered filter, or a raster object. The map will be resized to match the
dimensions of x.}

\item{x_channel, y_channel}{Which channel to use for the displacement in the
x and y direction.}

\item{scale}{How much displacement should a maximal channel value correspond
to? If a numeric it will be interpreted as pixel dimensions. If a unit object
it will be converted to pixel dimension when rendered.}

\item{...}{Arguments to be passed on to methods}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{background}{A grob to draw below the filtered grob.}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter displaces the pixels based on the colour values of another layer
or raster object. As such it can be used to distort the content of the layer.
}
\examples{
library(ggplot2)
ggplot() +
  as_reference(
    geom_polygon(aes(c(0, 1, 1), c(0, 0, 1)), colour = NA, fill = 'magenta' ),
    id = "displace_map"
  ) +
  with_displacement(
    geom_text(aes(0.5, 0.5, label = 'Displacements!'), size = 10),
    map = "displace_map",
    x_channel = "red",
    y_channel = "blue",
    scale = unit(0.025, 'npc')
  )

}
