% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow.R
\name{with_shadow}
\alias{with_shadow}
\title{Apply a drop shadow to a layer}
\usage{
with_shadow(layer, colour = "black", x_offset = 1, y_offset = 1,
  default_unit = "mm", radius = 0, sigma = 1, stack = TRUE)
}
\arguments{
\item{layer}{A ggplot2 layer object, usually constructed with a call to
\code{geom_*()} or \code{stat_*()}}

\item{colour}{The colour of the shadow}

\item{x_offset, y_offset}{The offset of the shadow from the origin}

\item{default_unit}{The unit of \code{x_offset} and \code{y_offset} if they are given
as numerics}

\item{radius}{The radius of the blur. If \code{0} it will be calculated
automatically from \code{sigma}}

\item{sigma}{The standard deviation of the gaussian kernel. Increase it to
apply more blurring}

\item{stack}{Should the original layer be placed on top? Defaults to \code{FALSE}}
}
\description{
This filter applies the familiar drop-shadow effect on elements in a layer.
It takes the outline of each shape, offsets it from its origin and applies a
blur to it.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_shadow(geom_point(colour = 'red', size = 3), sigma = 3)

}
