% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R, R/displace.R, R/kernel.R, R/kernels.R,
%   R/rasterise_grob.R
\name{blur_raster}
\alias{blur_raster}
\alias{displace_raster}
\alias{convolve_grob}
\alias{kernel_gaussian}
\alias{raster_helpers}
\title{Raster Helpers}
\usage{
blur_raster(x, sigma = 0.5, stack = FALSE)

displace_raster(x, map, x_channel, y_channel, scale)

convolve_grob(
  x,
  kernel,
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE
)

kernel_gaussian(sigma)
}
\description{
Raster Helpers
}
\keyword{internal}
