% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R, R/blur.R, R/displace.R,
%   R/interpolate.R, R/kernel.R, R/kernels.R, R/rasterise_grob.R
\name{blend_raster}
\alias{blend_raster}
\alias{blur_raster}
\alias{displace_raster}
\alias{interpolate_raster}
\alias{convolve_grob}
\alias{kernel_gaussian}
\alias{raster_helpers}
\title{Raster Helpers}
\usage{
blend_raster(x, bg_layer, blend_type = "Over", flip_order = FALSE)

blur_raster(x, sigma = 0.5, stack = FALSE)

displace_raster(x, map, x_channel, y_channel, scale)

interpolate_raster(x, bg_layer, src_percent, bg_percent)

convolve_grob(
  x,
  kernel,
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE
)

kernel_gaussian(sigma)
}
\description{
Raster Helpers
}
\keyword{internal}
