% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{with_blur}
\alias{with_blur}
\alias{with_blur.grob}
\alias{with_blur.Layer}
\alias{with_blur.ggplot}
\title{Apply a gaussian blur to your layer}
\usage{
with_blur(x, sigma = 0.5, stack = FALSE, ..., id = NULL, include = is.null(id))

\method{with_blur}{grob}(
  x,
  sigma,
  stack = FALSE,
  background = NULL,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_blur}{Layer}(x, sigma = 0.5, stack = FALSE, ..., id = NULL, include = is.null(id))

\method{with_blur}{ggplot}(
  x,
  sigma = 0.5,
  stack = FALSE,
  ignore_background = TRUE,
  ...,
  id = NULL,
  include = is.null(id)
)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{sigma}{The standard deviation of the gaussian kernel. Increase it to
apply more blurring. If a numeric it will be interpreted as given in pixels.
If a unit object it will automatically be converted to pixels at rendering
time}

\item{stack}{Should the original layer be placed on top?}

\item{...}{Arguments to be passed on to methods}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{background}{A grob to draw below the filtered grob.}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter adds a blur to the provided ggplot layer. The amount of blur can
be controlled and the result can optionally be put underneath the original
layer.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_blur(geom_point(data = mtcars, size = 3), sigma = 3)

}
