% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{with_blur}
\alias{with_blur}
\title{Apply a gaussian blur to your layer}
\usage{
with_blur(layer, sigma = 0.5, radius = 0, stack = FALSE)
}
\arguments{
\item{layer}{A ggplot2 layer object, usually constructed with a call to
\code{geom_*()} or \code{stat_*()}}

\item{sigma}{The standard deviation of the gaussian kernel. Increase it to
apply more blurring}

\item{radius}{The radius of the blur. If \code{0} it will be calculated
automatically from \code{sigma}}

\item{stack}{Should the original layer be placed on top? Defaults to \code{FALSE}}
}
\value{
A modified \code{Layer} object
}
\description{
This filter adds a blur to the provided ggplot layer. The amount of blur can
be controlled and the result can optionally be put underneath the original
layer.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_blur(geom_point(data = mtcars[4,], size = 3), sigma = 3)

}
