% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displace.R
\name{with_displacement}
\alias{with_displacement}
\alias{with_displacement.grob}
\alias{with_displacement.Layer}
\alias{with_displacement.ggplot}
\alias{with_displacement.character}
\alias{with_displacement.function}
\alias{with_displacement.formula}
\alias{with_displacement.element}
\alias{with_displacement.guide}
\title{Apply a displacement map to a layer}
\usage{
with_displacement(x, x_map, y_map = x_map, x_scale = 1, y_scale = x_scale, ...)

\method{with_displacement}{grob}(
  x,
  x_map,
  y_map = x_map,
  x_scale = 1,
  y_scale = x_scale,
  ...,
  background = NULL,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{Layer}(
  x,
  x_map,
  y_map = x_map,
  x_scale = 1,
  y_scale = x_scale,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{ggplot}(
  x,
  x_map,
  y_map = x_map,
  x_scale = 1,
  y_scale = x_scale,
  ignore_background = TRUE,
  ...
)

\method{with_displacement}{character}(
  x,
  x_map,
  y_map = x_map,
  x_scale = 1,
  y_scale = x_scale,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{`function`}(
  x,
  x_map,
  y_map = x_map,
  x_scale = 1,
  y_scale = x_scale,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{formula}(
  x,
  x_map,
  y_map = x_map,
  x_scale = 1,
  y_scale = x_scale,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_displacement}{element}(x, x_map, y_map = x_map, x_scale = 1, y_scale = x_scale, ...)

\method{with_displacement}{guide}(x, x_map, y_map = x_map, x_scale = 1, y_scale = x_scale, ...)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{x_map, y_map}{The displacement maps to use. Can either be a string
identifying a registered filter, or a raster object. The maps will be resized
to match the dimensions of x. Only one channel will be used - see
\link[=Channels]{the docs on channels} for info on how to set them.}

\item{x_scale, y_scale}{How much displacement should a maximal channel value
correspond to? If a numeric it will be interpreted as pixel dimensions. If a
unit object it will be converted to pixel dimension when rendered.}

\item{...}{Arguments to be passed on to methods}

\item{background}{A grob to draw below the filtered grob.}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter displaces the pixels based on the colour values of another layer
or raster object. As such it can be used to distort the content of the layer.
}
\examples{
library(ggplot2)
ggplot() +
  as_reference(
    geom_polygon(aes(c(0, 1, 1), c(0, 0, 1)), colour = NA, fill = 'magenta' ),
    id = "displace_map"
  ) +
  with_displacement(
    geom_text(aes(0.5, 0.5, label = 'Displacements!'), size = 10),
    x_map = ch_red("displace_map"),
    y_map = ch_blue("displace_map"),
    x_scale = unit(0.025, 'npc'),
    y_scale = unit(0.025, 'npc')
  )

}
