% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{with_interpolate}
\alias{with_interpolate}
\alias{with_interpolate.grob}
\alias{with_interpolate.Layer}
\alias{with_interpolate.ggplot}
\alias{with_interpolate.character}
\alias{with_interpolate.element}
\alias{with_interpolate.guide}
\title{Convert a layer to a raster}
\usage{
with_interpolate(x, bg_layer, src_percent, bg_percent = 100 - src_percent, ...)

\method{with_interpolate}{grob}(
  x,
  bg_layer,
  src_percent,
  bg_percent = 100 - src_percent,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_interpolate}{Layer}(
  x,
  bg_layer,
  src_percent,
  bg_percent = 100 - src_percent,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_interpolate}{ggplot}(
  x,
  bg_layer,
  src_percent,
  bg_percent = 100 - src_percent,
  ignore_background = TRUE,
  ...
)

\method{with_interpolate}{character}(
  x,
  bg_layer,
  src_percent,
  bg_percent = 100 - src_percent,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_interpolate}{element}(x, bg_layer, src_percent, bg_percent = 100 - src_percent, ...)

\method{with_interpolate}{guide}(x, bg_layer, src_percent, bg_percent = 100 - src_percent, ...)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{...}{Arguments to be passed on to methods}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter simply converts the given layer, grob, oor ggplot to a raster and
inserts it back again. It is useful for vector graphics devices such as
svglite if a layer contains a huge amount of primitives that would make the
file slow to render. \code{as_reference(x, id)} is a shorthand for
\code{with_interpolate(x, id = id, include = FALSE)} that makes the intent of using
this grob or layer as only a filter reference clear.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_interpolate(geom_point(data = mtcars, size = 3))

}
