% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{with_kernel}
\alias{with_kernel}
\alias{with_kernel.grob}
\alias{with_kernel.Layer}
\alias{with_kernel.ggplot}
\alias{with_kernel.character}
\alias{with_kernel.function}
\alias{with_kernel.formula}
\alias{with_kernel.element}
\alias{with_kernel.guide}
\title{Apply a gaussian blur to your layer}
\usage{
with_kernel(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...
)

\method{with_kernel}{grob}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  background = NULL,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_kernel}{Layer}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_kernel}{ggplot}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ignore_background = TRUE,
  ...
)

\method{with_kernel}{character}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_kernel}{`function`}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_kernel}{formula}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_kernel}{element}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...
)

\method{with_kernel}{guide}(
  x,
  kernel = kernel_gaussian(0.5),
  iterations = 1,
  scaling = NULL,
  bias = NULL,
  stack = FALSE,
  ...
)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{kernel}{either a square matrix or a string. The string can either be a
parameterized \link[magick:options]{kerneltype} such as: \code{"DoG:0,0,2"} or \code{"Diamond"}
or it can contain a custom matrix (see examples)}

\item{iterations}{number of iterations}

\item{scaling}{string with kernel scaling. The special flag \code{"!"} automatically scales to full
dynamic range, for example: \code{"50\%!"}}

\item{bias}{output bias string, for example \code{"50\%"}}

\item{stack}{Should the original layer be placed on top?}

\item{...}{Arguments to be passed on to methods}

\item{background}{A grob to draw below the filtered grob.}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter allows you to apply a custom kernel to your layer, thus giving
you more control than e.g. \code{\link[=with_blur]{with_blur()}} which is also applying a kernel.
}
\examples{
library(ggplot2)
# Add directional blur using the comet kernel
ggplot(mtcars, aes(mpg, disp)) +
  with_kernel(geom_point(size = 3), 'Comet:0,10')

}
