% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{with_kernel}
\alias{with_kernel}
\title{Apply a gaussian blur to your layer}
\usage{
with_kernel(layer, kernel = "Gaussian", iterations = 1,
  scaling = NULL, bias = NULL, stack = FALSE)
}
\arguments{
\item{layer}{A ggplot2 layer object, usually constructed with a call to
\code{geom_*()} or \code{stat_*()}}

\item{kernel}{either a square matrix or a string. The string can either be a
parameterized \link[=kernel_types]{kerneltype} such as: \code{"DoG:0,0,2"} or \code{"Diamond"}
or it can contain a custom matrix (see examples)}

\item{iterations}{number of iterations}

\item{scaling}{string with kernel scaling. The special flag \code{"!"} automatically scales to full
dynamic range, for example: \code{"50\%!"}}

\item{bias}{output bias string, for example \code{"50\%"}}

\item{stack}{Should the original layer be placed on top? Defaults to \code{FALSE}}
}
\value{
A modified \code{Layer} object
}
\description{
This filter allows you to apply a custom kernel to your layer, thus giving
you more control than e.g. \code{\link[=with_blur]{with_blur()}} which is also applying a kernel.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_blur(geom_point(data = mtcars[4,], size = 3), sigma = 3)

}
