% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{with_mask}
\alias{with_mask}
\alias{with_mask.grob}
\alias{with_mask.Layer}
\alias{with_mask.ggplot}
\alias{with_mask.character}
\alias{with_mask.function}
\alias{with_mask.formula}
\alias{with_mask.element}
\alias{with_mask.guide}
\title{Apply a mask to a layer}
\usage{
with_mask(x, mask, invert = FALSE, ...)

\method{with_mask}{grob}(
  x,
  mask,
  invert = FALSE,
  ...,
  background = NULL,
  id = NULL,
  include = is.null(id)
)

\method{with_mask}{Layer}(x, mask, invert = FALSE, ..., id = NULL, include = is.null(id))

\method{with_mask}{ggplot}(x, mask, invert = FALSE, ignore_background = TRUE, ...)

\method{with_mask}{character}(x, mask, invert = FALSE, ..., id = NULL, include = is.null(id))

\method{with_mask}{`function`}(x, mask, invert = FALSE, ..., id = NULL, include = is.null(id))

\method{with_mask}{formula}(x, mask, invert = FALSE, ..., id = NULL, include = is.null(id))

\method{with_mask}{element}(x, mask, invert = FALSE, ...)

\method{with_mask}{guide}(x, mask, invert = FALSE, ...)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{mask}{The layer to use as mask. Can either be a string
identifying a registered filter, or a raster object. Will by default extract
the luminosity of the layer and use that as mask. To pick another channel use
one of the \link[=Channels]{channel specification} function.}

\item{invert}{Should the mask bee}

\item{...}{Arguments to be passed on to methods}

\item{background}{A grob to draw below the filtered grob.}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter applies a mask to the given layer, i.e. sets the opacity of the
layer based on another layer
}
\examples{
library(ggplot2)
volcano_raster <- as.raster((volcano - min(volcano))/diff(range(volcano)))
circle <- data.frame(
  x = cos(seq(0, 2*pi, length.out = 360)),
  y = sin(seq(0, 2*pi, length.out = 360))
)

ggplot() +
  as_reference(
    geom_polygon(aes(x = x, y = y), circle),
    id = 'circle'
  ) +
  with_mask(
    annotation_raster(volcano_raster, -1, 1, -1, 1, TRUE),
    mask = ch_alpha('circle')
  )

# use invert = TRUE to flip the mask
ggplot() +
  as_reference(
    geom_polygon(aes(x = x, y = y), circle),
    id = 'circle'
  ) +
  with_mask(
    annotation_raster(volcano_raster, -1, 1, -1, 1, TRUE),
    mask = ch_alpha('circle'),
    invert = TRUE
  )

}
