% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{with_raster}
\alias{with_raster}
\alias{as_reference}
\alias{with_raster.grob}
\alias{with_raster.Layer}
\alias{with_raster.ggplot}
\title{Convert a layer to a raster}
\usage{
with_raster(x, ..., id = NULL, include = is.null(id))

as_reference(x, id)

\method{with_raster}{grob}(x, ..., id = NULL, include = is.null(id))

\method{with_raster}{Layer}(x, ..., id = NULL, include = is.null(id))

\method{with_raster}{ggplot}(x, ignore_background = TRUE, ..., id = NULL, include = is.null(id))
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{...}{Arguments to be passed on to methods}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\value{
A modified \code{Layer} object
}
\description{
This filter simply converts the given layer, grob, oor ggplot to a raster and
inserts it back again. It is useful for vector graphics devices such as
svglite if a layer contains a huge amount of primitives that would make the
file slow to render. \code{as_reference(x, id)} is a shorthand for
\code{with_raster(x, id = id, include = FALSE)} that makes the intent of using
this grob or layer as only a filter reference clear.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_raster(geom_point(data = mtcars, size = 3))

}
