% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow.R
\name{with_shadow}
\alias{with_shadow}
\alias{with_shadow.grob}
\alias{with_shadow.Layer}
\alias{with_shadow.ggplot}
\title{Apply a drop shadow to a layer}
\usage{
with_shadow(
  x,
  colour = "black",
  x_offset = 1,
  y_offset = 1,
  default_unit = "mm",
  sigma = 1,
  stack = TRUE,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_shadow}{grob}(
  x,
  colour = "black",
  x_offset = 1,
  y_offset = 1,
  default_unit = "mm",
  sigma = 1,
  stack = TRUE,
  background = NULL,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_shadow}{Layer}(
  x,
  colour = "black",
  x_offset = 1,
  y_offset = 1,
  default_unit = "mm",
  sigma = 1,
  stack = TRUE,
  ...,
  id = NULL,
  include = is.null(id)
)

\method{with_shadow}{ggplot}(
  x,
  colour = "black",
  x_offset = 1,
  y_offset = 1,
  default_unit = "mm",
  sigma = 1,
  stack = TRUE,
  ignore_background = TRUE,
  ...,
  id = NULL,
  include = is.null(id)
)
}
\arguments{
\item{x}{A ggplot2 layer object, a ggplot, a grob, or a character string
naming a filter}

\item{colour}{The colour of the shadow}

\item{x_offset, y_offset}{The offset of the shadow from the origin}

\item{default_unit}{The unit of \code{x_offset} and \code{y_offset} if they are given
as numerics}

\item{sigma}{The standard deviation of the gaussian kernel. Increase it to
apply more blurring. If a numeric it will be interpreted as given in pixels.
If a unit object it will automatically be converted to pixels at rendering
time}

\item{stack}{Should the original layer be placed on top?}

\item{...}{Arguments to be passed on to methods}

\item{id}{An id that can be used to reference this filter somewhere else}

\item{include}{Should the filter be part of the final render}

\item{background}{A grob to draw below the filtered grob.}

\item{ignore_background}{Should the filter be applied to everything except
the plot background, or should the background be included.}
}
\description{
This filter applies the familiar drop-shadow effect on elements in a layer.
It takes the outline of each shape, offsets it from its origin and applies a
blur to it.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  with_shadow(geom_point(colour = 'red', size = 3), sigma = 3)

}
